backend_merge/bot/requirements.txt
Flask
Flask-SQLAlchemy
twilio
python-dotenv
web3 # optional, only when implementing chain
requests # For API requests
pytest
pytest-cov
typeguard
anyio
openai-whisper
gTTS
torch
ffmpeg-python # For audio format conversion (optional, uses system ffmpeg)